SE-36: Add worflow of deploying API docs
testing

add permission

Publish `API` dir

Add swagger ui dist

Create docs in workflow

Implement operations sorter
