+++ api/package.json
@@ -6,6 +6,7 @@
     "@": "dist"
   },
   "scripts": {
+    "generate-swagger": "tsc && node -r module-alias/register dist/swagger/generate-swagger.js",
     "lint": "eslint \"./**/*.ts\"",
     "prettier": "prettier --check \"./**/*.{ts,json}\"",
     "prettier-fix": "prettier --write \"./**/*.{ts,json}\"",
@@ -21,6 +22,7 @@
     "sharp": "^0.33.4",
     "streamifier": "^0.1.1",
     "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-dist": "^5.17.14",
     "swagger-ui-express": "^5.0.1",
     "uuid": "^10.0.0"
   },
