+++ api/src/swagger/generate-swagger.ts
@@ -1,0 +1,4 @@
+import { swaggerSpec } from ".";
+import fs from "fs";
+
+fs.writeFileSync("swagger.json", JSON.stringify(swaggerSpec, null, 2));
