+++ api/src/swagger/index.ts
@@ -1,7 +1,9 @@
 import swaggerJSDoc from "swagger-jsdoc";
 import swaggerUi, { SwaggerUiOptions } from "swagger-ui-express";
 import { Router } from "express";
-import config from "./config";
+
+import config from "../config";
+import operationsSorter from "./operations-sorter";
 
 const options = {
   definition: {
@@ -34,26 +36,7 @@ const swaggerSpec = swaggerJSDoc(options);
 
 const swaggerOptions: SwaggerUiOptions = {
   swaggerOptions: {
-    operationsSorter: (a: any, b: any) => {
-      const methodsOrder = [
-        "get",
-        "post",
-        "put",
-        "patch",
-        "delete",
-        "options",
-        "trace",
-      ];
-      let result =
-        methodsOrder.indexOf(a.get("method")) -
-        methodsOrder.indexOf(b.get("method"));
-
-      if (result === 0) {
-        result = a.get("path").localeCompare(b.get("path"));
-      }
-
-      return result;
-    },
+    operationsSorter,
   },
 };
 
@@ -67,4 +50,5 @@ const swaggerRouter = Router()
     res.json(swaggerSpec);
   });
 
+export { swaggerSpec };
 export default swaggerRouter;
